+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated GitHub CI to use Windows 2025 for Windows builds.
 - Significant rework of the GitHub CI workflow for improved reliability and maintainability.
 - Updated Prettier to the latest version.
+- Use shared library suffix CMAKE macro.
 
 ## 1.13.0 - (2025-03-29)